feat: added access token update in accept invite. docs: fixed date and invite docs fix: fix delete date route refactor: transformed CreateTokens in 2 separeted methods fix:  changed id icon, id relationship and id author types in db

+++ src/main/resources/db/migration/V4__create_dates_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE dates(
     local VARCHAR(250) NOT NULL,
     title VARCHAR(500) NOT NULL,
     description VARCHAR(500) NOT NULL,
-    id_icon INTEGER NOT NULL,
-    id_relationship INTEGER NOT NULL,
-    id_author INTEGER NOT NULL
+    id_icon BIGINT NOT NULL,
+    id_relationship BIGINT NOT NULL,
+    id_author BIGINT NOT NULL
 );
